Add class handling upload.
