+++ includes/Jobs/JobRegistry.php
@@ -4,6 +4,7 @@ namespace SkyVerge\WooCommerce\Facebook\Jobs;
 
 use Automattic\WooCommerce\ActionSchedulerJobFramework\Proxies\ActionScheduler;
 use SkyVerge\WooCommerce\Facebook\Feed\FeedFileHandler;
+use SkyVerge\WooCommerce\Facebook\Feed\FeedFileUploadHandler;
 use SkyVerge\WooCommerce\Facebook\Feed\FeedDataExporter;
 
 
@@ -30,6 +31,7 @@ class JobRegistry {
 		$feed_data_exporter              = new FeedDataExporter();
 		$this->generate_product_feed_job = new GenerateProductFeed( $action_scheduler_proxy, $feed_file_handler, $feed_data_exporter );
 		$this->generate_product_feed_job->init();
+		FeedFileUploadHandler::init( $feed_file_handler );
 	}
 
 }
